test: restore apply_socket_options coverage as inline module
apply_socket_options had no test coverage: the tests written for it lived
in crates/core/src/client/tests/module_list_socket_options.rs, which was
orphaned when client/mod.rs dropped `mod tests;` (commit 780ff004a), so
they stopped compiling and running.

Relocate the three self-contained tests into an inline #[cfg(test)] module
in socket_options/apply.rs (the idiomatic Rust pattern) rather than
re-wiring the fragile include! chain, and delete the orphaned file. They
cover send-buffer sizing, warn-and-continue on an unknown option name, and
an OPT_ON option given a value - mirroring upstream socket.c's
warn-and-continue set_socket_options() contract.
