+++ crates/core/src/client/tests/mod.rs
@@ -7,6 +7,5 @@ include!("module_list_request.rs");
 include!("module_list_proxy.rs");
 include!("module_list_auth.rs");
 include!("module_list_listing.rs");
-include!("module_list_socket_options.rs");
 include!("error_helpers.rs");
 include!("ignore_times.rs");
